Moves some modules and classes to under the abstract directory.
