+++ activerecord/lib/active_record/connection_adapters/abstract/column.rb
