+++ activerecord/lib/active_record/connection_adapters/abstract/connection_specification.rb
