+++ activerecord/lib/active_record/connection_adapters/abstract/schema_cache.rb
