+++ activerecord/lib/active_record/connection_adapters/abstract/statement_pool.rb
