+++ activerecord/lib/active_record/connection_adapters/mysql_adapter.rb
@@ -1,5 +1,4 @@
 require 'active_record/connection_adapters/abstract_mysql_adapter'
-require 'active_record/connection_adapters/statement_pool'
 require 'active_support/core_ext/hash/keys'
 
 gem 'mysql', '~> 2.9'
