+++ activerecord/lib/active_record/connection_adapters/sqlite3_adapter.rb
@@ -1,5 +1,4 @@
 require 'active_record/connection_adapters/abstract_adapter'
-require 'active_record/connection_adapters/statement_pool'
 require 'arel/visitors/bind_visitor'
 
 gem 'sqlite3', '~> 1.3.6'
